+++ Beam/resources/views/articles/pageviews.blade.php
@@ -83,6 +83,13 @@
                     'priority' => 5,
                     'className' => 'text-right',
                 ],
+                'content_type' => [
+                    'header' => 'Type',
+                    'orderable' => false,
+                    'filter' => $contentTypes,
+                    'priority' => 2,
+                    'visible' => count($contentTypes) > 1,
+                ],
                 'authors' => [
                     'header' => 'authors',
                     'orderable' => false,